requirements.txt
asttokens==2.4.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
idna==3.6
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
langdetect==1.0.9
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
numpy==1.26.4
parso==0.8.3
playsound==1.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
setuptools==65.5.1
six==1.16.0
sounddevice==0.4.6
SpeechRecognition==3.10.1
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing-extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.38.4
